email-rag/requirements.txt
streamlit==1.35.0
imapclient==3.0.1
email-validator==2.1.1
beautifulsoup4==4.12.3
chromadb==0.5.0
langchain==0.2.1
langchain-community==0.2.1
schedule==1.2.1
python-dotenv==1.0.1
msal==1.28.0
lxml==5.2.2
html2text==2024.2.26